Refactor Components
1. Move Header and NavBar Styles to components/header/
2. Added Responsive Styles for Nav - 900px or greater
Note: Plan to add new component for Mobile NavBar
